assign n grp create bug fix
