+++ lms/templates/assignments/home_assignment.html
@@ -15,7 +15,7 @@
                             </small>
                         </div>
                         <h2><a class="article-title"
-                               href="{% url 'lms:assignment_student_detail' object.courses.id student_assignment.assignment.id %}">
+                               href="{% url 'lms:assignment_student_detail' object1.id student_assignment.assignment.id %}">
                             {{ student_assignment.assignment.name }}
                         </a>
                         </h2>
