+++ lms/templates/lms/course/course_home.html
@@ -11,7 +11,7 @@
     <h1 class='text-center'>{{ object.title }}</h1>
     <article class="media content-section">
         <div class="media-body">
-  
+
           <p class="article-content">{{ object.description }}</p>
         </div>
     </article>
