+++ lms/templates/lms/course/view_groups.html
@@ -10,9 +10,11 @@
         <h1 class="text-center pt-3 pb-3 border-bottom">View Groups</h1>
 
         <div class="mb-lg-5">
-            <p>Click here to create a group:</p>
-            <a class="btn btn-outline-primary btn-sm mt-1 mb-1" href="{% url 'lms:group_creation_request' object.id %}">
-                Create Group</a>
+            {% if request.user.role.is_student %}
+              <p>Click here to create a group:</p>
+              <a class="btn btn-outline-primary btn-sm mt-1 mb-1" href="{% url 'lms:group_creation_request' object.id %}">
+                  Create Group</a>
+            {% endif %}      
         </div>
 
         <div class="row">
