+++ lms/templates/quiz/compute_stats.html
