+++ lms/templates/quiz/quiz_preview.html
