+++ lms/views/course/course_views.py
@@ -14,7 +14,6 @@ from lms.models.course_model import Course , StudentCourse
 from lms.tables import StudentAssignmentTable, StudentAssignmentFilter
 from lms.models.user_role_model import Role
 
-from lms.models.course_model import Question, Responses, Quiz, Student_Question, UsersProfile
 from django.db.models import Avg, Max, Min
 from django.core.paginator import Paginator
 
@@ -23,32 +22,24 @@ class CourseListView(ListView):
     model = Course
     context_object_name = "courses"
     context={}
-    template_name = "lms/course/home.html"
-   
-
-    # def get(self, request, *args, **kwargs):
-    #     selected_course = Course.objects.get(id=kwargs['pk'])
-    #     print(f'SELECTED COURSE {selected_course}')
-    #     # selected_course.published=False
-    #     # selected_course.save()
-    #     return redirect('lms:dashboard_profile')
+    template_name = "lms/course/course_home.html"
 
     def get(self, request, *args, **kwargs):
-        
+
         role = Role.objects.filter(user=request.user)
         if role[0].is_student:
             selected_student_course = StudentCourse.objects.get(id=kwargs['pk'])
+            print(selected_student_course)
             selected_course = Course.objects.get(id=selected_student_course.courses_id)
+            print(selected_course)
             self.context.update(object1=selected_student_course)
             self.context.update(object=selected_course)
         else:
             selected_course = Course.objects.get(id=kwargs['pk'])
             self.context.update(object=selected_course)
 
-    # self.context.update(course_publish=selected_course.published)
-        # self.context.update(course_description=selected_course.description)
         return render(request, self.template_name, self.context)
-        
+
 
 class CourseListView_default(ListView):
     model = Course
@@ -93,106 +84,3 @@ def table_download(request):
     return render(request, "lms/course/gradebook/course_gradebook_export.html", {
         "table": table
     })
-
-
-# Fetch Quiz Questions all at a time by Quiz Teacher View Team.
-def fetch_questions(request):
-    questions = Question.objects.all()
-    quizs = Quiz.objects.first()
-    context = dict()
-    context.update({'questions':questions})
-    context.update({'quizs':quizs})
-    return render(request,'lms/course/quiz3.html',context)
-
-
-# Fetch one quiz question at a time.
-def fetch_questions_oneatatime(request):
-    obj = Question.objects.all()
-
-    count = Question.objects.all().count()
-    paginator = Paginator(obj,1)
-    try:
-        page = int(request.GET.get('page','1'))  
-    except:
-        page =1
-    try:
-        questions = paginator.page(page)
-    except(EmptyPage,InvalidPage):
-
-        questions=paginator.page(paginator.num_pages)
-    
-    return render(request,'lms/course/quiz.html',{'obj':obj,'questions':questions,'count':count})    
-
-
-# Quiz Starting Page.
-def display_lp(request):
-    quizs = Quiz.objects.first()
-    context = {'quizs':quizs}
-    return render(request,'lms/quiz2.html',context)
-
-
-# Edit Quiz Questions only admin has the right to to do.
-def Edit_quiz(request):
-    return render(request,'admin/')  
-
-
-# Preview Quiz
-def preview_quiz(request):
-    questions = Question.objects.all()
-    context = {'questions':questions}
-    return render(request,'lms/course/quiz_preview.html',context)
-
-
-# Quiz Landing Page
-def quiz_lp(request):
-    quizs=Quiz.objects.all()
-    context = {'quizs':quizs}
-    return render(request,'lms/course/quizlp.html',context)
-
-
-# Compute Stats of Quiz
-def compute_stats(request):
-    profiles= UsersProfile.objects.all()
-    context= dict()
-    context.update({'profiles':profiles})
-    context.update(UsersProfile.objects.all().aggregate(Avg('marks')))
-    context.update(UsersProfile.objects.all().aggregate(Min('marks')))
-    context.update(UsersProfile.objects.all().aggregate(Max('marks')))
-    return render(request,'lms/course/compute_stats.html',context)
-
-
-# Enter Comments.
-def enter_comment(request):
-    responses = Responses.objects.all()
-    context = {'responses':responses}
-    if request.method == "POST":
-       responses = Responses()
-       responses.comments = request.POST['comments']
-       responses.save()
-       return render(request,'lms/course/responses.html')
-    else:
-        return render(request,'lms/course/responses.html',context)
-
-
-
-# Quiz Publish
-def quiz_publish(request):
-    questions = Question.objects.all()
-    for i in questions:
-        model_one = Student_Question(course= i.course,
-                                     quiznumber = i.quiznumber,
-                                     questiontype= i.questiontype,
-                                     questionnumber = i.questionnumber,
-                                     question= i.question,
-                                     description= i.description,
-                                     answer= i.answer,
-                                     option1= i.option1,
-                                     option2= i.option2,
-                                     option3= i.option3,
-                                     option4= i.option4,
-                                     points = i.points,
-                                     img= i.img,
-                                     files= i.files 
-                                     )
-        model_one.save()
-    return render(request,'lms/quiz_student_view.html')
